+++ ccvm_simulators/solution.py
@@ -18,6 +18,9 @@ class Solution:
         solve_time (float): Time to solve the problem.
         pp_time (float):  Time to post-process the problem.
         optimal_value (float): The optimal objective value for the given problem instance.
+        best_value (float): The best objective value for the given problem instance.
+        num_frac_values (int): Number of fractional values in the solutions of given problem instance.
+        solution_vector (list): Vector of solution to the problem instance found using Gurobi.
         variables (dict): A dict object to store solution related variables.
             Some fields might be solver-dependent, such as mu, sigma and s.
             - problem_variables (torch.Tensor): The values of the problem
@@ -50,6 +53,9 @@ class Solution:
     solve_time: float
     pp_time: float
     optimal_value: float
+    best_value: float
+    num_frac_values: int
+    solution_vector: list
     variables: dict = field(repr=False)
     evolution_file: str = None
     device: str = field(default="cpu", repr=False)