+++ ccvm_simulators/solvers/dl_solver.py
@@ -409,6 +409,8 @@ class DLSolver(CCVMSolver):
         confs = self.change_variables(problem_variables, S)
         objval = instance.compute_energy(confs)
 
+        # TODO: Move the rest of the code to DLSolver.__call__() and update the return accordingly
+
         if evolution_step_size:
             # Write samples to file
             # Overwrite file if it exists
@@ -433,6 +435,9 @@ class DLSolver(CCVMSolver):
             solve_time=solve_time,
             pp_time=pp_time,
             optimal_value=instance.optimal_sol,
+            best_value=instance.best_sol,
+            num_frac_values=instance.num_frac_values,
+            solution_vector=instance.solution_vector,
             variables={
                 "problem_variables": problem_variables,
                 "s": s,
@@ -651,7 +656,6 @@ class DLSolver(CCVMSolver):
                 v_c = beta2 * v_c + (1.0 - beta2) * torch.pow(c_grads, 2)
                 v_s = beta2 * v_s + (1.0 - beta2) * torch.pow(s_grads, 2)
 
-
                 # Compute bias correction in 2nd moment
                 beta2i = 1.0 - beta2 ** (i + 1)
                 vhat_c = v_c / beta2i
@@ -729,6 +733,8 @@ class DLSolver(CCVMSolver):
         confs = self.change_variables(problem_variables, S)
         objval = instance.compute_energy(confs)
 
+        # TODO: Move the rest of the code to DLSolver.__call__() and update the return accordingly
+
         if evolution_step_size:
             # Write samples to file
             # Overwrite file if it exists
@@ -753,6 +759,9 @@ class DLSolver(CCVMSolver):
             solve_time=solve_time,
             pp_time=pp_time,
             optimal_value=instance.optimal_sol,
+            best_value=instance.best_sol,
+            num_frac_values=instance.num_frac_values,
+            solution_vector=instance.solution_vector,
             variables={
                 "problem_variables": problem_variables,
                 "s": s,
@@ -767,14 +776,14 @@ class DLSolver(CCVMSolver):
         return solution
 
     def __call__(
-            self,
-            instance,
-            post_processor=None,
-            pump_rate_flag=True,
-            g=0.05,
-            evolution_step_size=None,
-            evolution_file=None,
-            algorithm_parameters=None,
+        self,
+        instance,
+        post_processor=None,
+        pump_rate_flag=True,
+        g=0.05,
+        evolution_step_size=None,
+        evolution_file=None,
+        algorithm_parameters=None,
     ):
         """Solves the given problem instance choosing one of the available DL-CCVM solvers.
 
@@ -802,6 +811,9 @@ class DLSolver(CCVMSolver):
             solution (Solution): The solution to the problem instance.
         """
 
+        # TODO: Get the solution object from DLSolver._solve() or _solve_adam()
+        #      and return the outcomes properly
+
         if algorithm_parameters is None:
             # Use the original DL solver
             return self._solve(
@@ -810,7 +822,7 @@ class DLSolver(CCVMSolver):
                 pump_rate_flag,
                 g,
                 evolution_step_size,
-                evolution_file
+                evolution_file,
             )
         elif isinstance(algorithm_parameters, AdamParameters):
             # Use the DL solver with the Adam algorithm
@@ -821,7 +833,9 @@ class DLSolver(CCVMSolver):
                 pump_rate_flag,
                 g,
                 evolution_step_size,
-                evolution_file
+                evolution_file,
             )
         else:
-            raise ValueError(f"Solver option type {type(algorithm_parameters)} is not supported.")
+            raise ValueError(
+                f"Solver option type {type(algorithm_parameters)} is not supported."
+            )