+++ ccvm_simulators/solvers/langevin_solver.py
@@ -340,6 +340,9 @@ class LangevinSolver(CCVMSolver):
         # Stop the timer for the solve
         solve_time = time.time() - solve_time_start
 
+        # TODO: Similar to DLSolver, move the rest of the code to the LangevinSolver.__call__()
+        #       and update the return accordingly
+
         # Run the post processor on the results, if specified
         if post_processor:
             post_processor_object = PostProcessorFactory.create_postprocessor(
@@ -380,6 +383,9 @@ class LangevinSolver(CCVMSolver):
             solve_time=solve_time,
             pp_time=pp_time,
             optimal_value=instance.optimal_sol,
+            best_value=instance.best_sol,
+            num_frac_values=instance.num_frac_values,
+            solution_vector=instance.solution_vector,
             variables={"problem_variables": problem_variables},
             device=device,
         )
@@ -574,6 +580,10 @@ class LangevinSolver(CCVMSolver):
         # Stop the timer for the solve
         solve_time = time.time() - solve_time_start
 
+        # TODO: Similar in DLSolver._solve() and DLSolver._solve_adam(),
+        #       move the rest of the code segments to the LangevinSolver.__call__()
+        #       and update the return accordingly
+
         # Run the post processor on the results, if specified
         if post_processor:
             post_processor_object = PostProcessorFactory.create_postprocessor(
@@ -614,6 +624,9 @@ class LangevinSolver(CCVMSolver):
             solve_time=solve_time,
             pp_time=pp_time,
             optimal_value=instance.optimal_sol,
+            best_value=instance.best_sol,
+            num_frac_values=instance.num_frac_values,
+            solution_vector=instance.solution_vector,
             variables={"problem_variables": problem_variables},
             device=device,
         )
@@ -657,9 +670,19 @@ class LangevinSolver(CCVMSolver):
         """
         if algorithm_parameters is None:
             # Use the original Langevin solver
-            return self._solve(instance, post_processor, evolution_step_size, evolution_file)
+            return self._solve(
+                instance, post_processor, evolution_step_size, evolution_file
+            )
         elif isinstance(algorithm_parameters, AdamParameters):
             # Use the Langevin solver with the Adam algorithm
-            return self._solve_adam(instance, algorithm_parameters.to_dict(), post_processor, evolution_step_size, evolution_file)
+            return self._solve_adam(
+                instance,
+                algorithm_parameters.to_dict(),
+                post_processor,
+                evolution_step_size,
+                evolution_file,
+            )
         else:
-            raise ValueError(f"Solver option type {type(algorithm_parameters)} is not supported.")
+            raise ValueError(
+                f"Solver option type {type(algorithm_parameters)} is not supported."
+            )