+++ ccvm_simulators/tests/test_solution.py
@@ -5,7 +5,7 @@ import os
 
 
 class TestSolution(unittest.TestCase):
-    def setUp(self):
+    def setUp(self):       
         self.problem_size = 10
         self.batch_size = 5
         self.instance_name = "test_instance"
@@ -13,6 +13,9 @@ class TestSolution(unittest.TestCase):
         self.solve_time = 2.0
         self.pp_time = 3.0
         self.optimal_value = 3.2
+        self.best_value = 3.2
+        self.num_frac_values = 1
+        self.solution_vector = [1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 1.0]
         self.device = "cpu"
         self.evolution_file = "test"
         self.iterations = 1
@@ -28,6 +31,9 @@ class TestSolution(unittest.TestCase):
             self.solve_time,
             self.pp_time,
             self.optimal_value,
+            self.best_value,
+            self.num_frac_values,
+            self.solution_vector,
             self.variables,
             self.evolution_file,
             self.device,
@@ -49,6 +55,9 @@ class TestSolution(unittest.TestCase):
             self.solve_time,
             self.pp_time,
             self.optimal_value,
+            self.best_value,
+            self.num_frac_values,
+            self.solution_vector,
             variable,
             self.evolution_file,
             self.device,
@@ -71,6 +80,9 @@ class TestSolution(unittest.TestCase):
             self.solve_time,
             self.pp_time,
             self.optimal_value,
+            self.best_value,
+            self.num_frac_values,
+            self.solution_vector,
             self.variables,
             self.evolution_file,
             self.device,
@@ -102,6 +114,9 @@ class TestSolution(unittest.TestCase):
             self.solve_time,
             self.pp_time,
             self.optimal_value,
+            self.best_value,
+            self.num_frac_values,
+            self.solution_vector,
             self.variables,
             self.evolution_file,
             self.device,
@@ -135,6 +150,9 @@ class TestSolution(unittest.TestCase):
             self.solve_time,
             self.pp_time,
             self.optimal_value,
+            self.best_value,
+            self.num_frac_values,
+            self.solution_vector,
             self.variables,
             self.evolution_file,
             self.device,
@@ -155,8 +173,7 @@ class TestSolution(unittest.TestCase):
         assert original_solution_stats == expected_solution_stats
 
     def test_get_meta_dict(self):
-        """Test solution performance for values parameters with no impact on solution stats"""
-
+        """Test solution performance for values parameters with no impact on solution stats"""    
         solution = Solution(
             self.problem_size,
             self.batch_size,
@@ -166,6 +183,9 @@ class TestSolution(unittest.TestCase):
             self.solve_time,
             self.pp_time,
             self.optimal_value,
+            self.best_value,
+            self.num_frac_values,
+            self.solution_vector,
             self.variables,
             self.evolution_file,
             self.device,
@@ -179,7 +199,10 @@ class TestSolution(unittest.TestCase):
             "solve_time": self.solve_time,
             "pp_time": self.pp_time,
             "optimal_value": self.optimal_value,
+            "best_value": self.best_value,
+            "num_frac_values": self.num_frac_values,
             "evolution_file": self.evolution_file,
+            "solution_vector": self.solution_vector,
             "solution_performance": {
                 "optimal": 0.0,
                 "one_percent": 0.0,